requirements.txt
Django
django-filter